Refactor layouts DashboardLayout and FullPageLayout MU-17
