+++ src/src/App.js
@@ -1,21 +1,24 @@
+import { CssBaseline } from "@mui/material";
+import { ThemeProvider } from "@mui/material/styles";
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Routes, Route, Navigate, useLocation } from "react-router-dom";
-import { ThemeProvider } from "@mui/material/styles";
-import { CssBaseline } from "@mui/material";
+
+import { Navigate, Route, Routes, useLocation } from "react-router-dom";
 
 import theme from "assets/theme";
+
 import routes from "routes";
 
-import { getAllCategory, getPermissions } from "redux/groupsCategory/slice";
 import {
   allCategoriesSelector,
   getCategoryPermissionsSelector
 } from "redux/groupsCategory/selector";
-import { getCurrentUserSlice } from "redux/currentUser/slice";
+import { getAllCategory, getPermissions } from "redux/groupsCategory/slice";
+
 import { getCurrentUserSelector } from "redux/currentUser/selector";
 import { isExpiredToken } from "utils";
 
+import { getCurrentUserSlice } from "redux/currentUser/slice";
 import "./index.css";
 
 function App() {
