+++ src/src/pages/GroupDetail/index.js
@@ -4,7 +4,7 @@ import { Card, Grid } from "@mui/material";
 import { useSelector, useDispatch } from "react-redux";
 import "./style.css";
 
-import DashboardLayout from "layouts/LayoutContainers/DashboardLayout";
+import DashboardLayout from "layouts/DashboardLayout";
 import DashboardNavbar from "layouts/Navbars/DashboardNavbar";
 
 import MDBox from "components/MDComponents/MDBox";
